trying to resolve bug before I continue

+++ task/src/main/java/com/enviro/assessmnet/grade001/tsolokhambule/task/model/WasteCategoryNotFoundException.java
@@ -1,7 +1,7 @@
 package com.enviro.assessmnet.grade001.tsolokhambule.task.model;
 
 public class WasteCategoryNotFoundException extends RuntimeException{
-     public WasteCategoryNotFoundException(Integer id){
+     public WasteCategoryNotFoundException(Long id){
         super("Could not find category with this id " + id);
     }
 }
