+++ task/src/main/java/com/enviro/assessmnet/grade001/tsolokhambule/task/model/WasteCategoryRepository.java
@@ -2,7 +2,7 @@ package com.enviro.assessmnet.grade001.tsolokhambule.task.model;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WasteCategoryRepository extends JpaRepository<WasteCategory, Integer>{
+public interface WasteCategoryRepository extends JpaRepository<WasteCategory, Long>{
 
     
 }
